Refactor structure of auth dir
