+++ backend/auth/__init__.py
@@ -1,0 +1,3 @@
+from flask import Blueprint
+
+auth = Blueprint('auth', __name__)
